requirements.txt
coverage==7.6.12
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.0.1
